_archive/requirements.txt
Pillow
reportlab
PyQt6
